feat(核心功能): 全面優化車輛移動控制與自動駕駛模式
### **變更摘要**
本次提交專注於解決車輛在手動與自動模式下的核心問題，並提升了系統的整體穩定性與反應速度。

### **詳細變更**
- **韌體 (Arduino & ESP32)**
  - `arduino_uno.ino`:
    - 調整了主控制迴圈的邏輯，優化了感測器數據的讀取與處理流程。
    - 改善了自動避障模式的演算法，使其反應更為靈敏、路徑決策更佳。
  - `esp32_cam.ino`:
    - 調整影像傳輸參數，降低串流延遲，提升即時影像品質。
  - `MotorController/`:
    - 修正 `MotorController.cpp` 中導致馬達輸出不穩定的計算錯誤。
    - 在 `VehicleData.h` 中新增了狀態旗標，用於更精確地同步車輛狀態。

- **後端 (Python)**
  - `apis/vehicle_api.py`:
    - 更新 API 端點以匹配韌體的變更，特別是手動控制與模式切換的邏輯。
  - `services/camera_stream_processor.py`:
    - 增強了影像處理服務，以適應來自 ESP32 的新影像格式，並改善了物件偵測的初步邏輯。

- **前端 (HTML)**
  - `templates/index.html`:
    - 更新了前端控制介面，新增了模式狀態顯示，並使手動控制按鈕反應更即時。

### **測試**
- 已通過手動控制測試 (`test_manual_control.sh`)。
- 已通過自動避障模式場景測試 (`test_set_control_mode_avoidance.sh`)。
- 後端 API 已通過單元測試。

+++ src/miniauto/libraries/MotorController/MotorController.h
@@ -12,7 +12,7 @@ public:
   void stop();
 
 private:
-  void setMotors(int8_t m0, int8_t m1, int8_t m2, int8_t m3);
+  void setMotors(int16_t m0, int16_t m1, int16_t m2, int16_t m3);
   const static uint8_t pwm_min = 2;
   const static uint8_t motorpwmPin[4];
   const static uint8_t motordirectionPin[4];
